Disallow authentication of users with a status other than enabled
Prior to this commit, any user with a valid username and password
would be authenticated. With this commit, users must also have a
status of enabled in order to authenticate. This is a quick fix to a
problem specific to a particular use of rubycas-server. A more general
solution would be needed before this change should be considered for
merging upstream.