[PE-977] Check service url to prevent phishing
Add check_service_url method to check the service url against the console base
url or an optional service whitelist based on scheme, host, and port. If there
is no service param given or if checking the service param is disabled, we
should skip over all the checking and just return the cleaned service url.

Add tests for service url checking. Test what happens if a service does not
match the base url or a service on the whitelist, if a service url does match
a service on the whitelist, and if service checking is turned off.