(PE-1364) (#22317) configify rubycas-server/config.yml
Currently /etc/puppetlabs/rubycas-server/config.yml is just an empty file laid
down by the pe-rubycas-server package. The upgrader has logic to ensure that it
only doesn't replace this file if it has content, but if its empty just
overwrites it. This is insufficient though, because we also install the
rubycas-server package during upgrades prior to the upgrader logic. Since its
not marked as a config file, the package gladly overwrites it with a blank
file.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>